+++ src/components/Inputs/ActionAppBar/ActionAppBar.tsx
@@ -6,6 +6,10 @@ const ActionAppBar = ({ sx, ...props }: ButtonProps) => (
     variant="outlined"
     size="small"
     sx={{
+      "&:hover": {
+        backgroundColor: dark?.palette?.action?.hover,
+        borderColor: dark?.palette?.primary.main,
+      },
       backgroundColor: dark?.palette?.grey?.["100"],
       borderColor: dark?.palette?.divider,
       color: dark?.palette?.text?.primary,