+++ src/constants/colors.ts
@@ -1,6 +1,4 @@
-import { ThemeOptions } from "@mui/material";
-
-const light: ThemeOptions = {
+const light = {
   palette: {
     action: {
       active: "rgba(0, 0, 0, 0.54)",
@@ -131,7 +129,7 @@ const light: ThemeOptions = {
   },
 };
 
-const dark: ThemeOptions = {
+const dark = {
   palette: {
     action: {
       active: "rgba(255, 255, 255, 0.54)",