+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.0.0-beta.8
-- **[fix]** - TextFieldAppBar InputProps
+## v2.0.0-beta.9
+- **[fix]** - Tab lowercase
+- **[fix]** - TextFieldAppBar light mode
+- **[fix]** - ActionAppBar light mode
+- **[fix]** - Letter spacing